interim

+++ mxmd/src/MxMDFrame.hpp
@@ -40,7 +40,7 @@ struct MxMDFrame {
   uint16_t	type = 0;
   uint64_t	linkID = 0;	// shard ID when used for ITC/IPC
   uint64_t  	seqNo = 0;
-  int32_t	sec = 0;
+  int64_t	sec = 0;
   uint32_t	nsec = 0;
 };
